Put useful utils into @pigi/utils
