+++ packages/core/src/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './misc'
 export * from './range-store'
 export * from './state-manager'
